fix checks for cache

+++ inadyn/rootfs/etc/cont-init.d/99-run.sh
@@ -6,17 +6,26 @@
 #################
 mustache-cli /data/options.json /templates/inadyn.mustache >/etc/inadyn.conf
 
+# If cache folder not exist then make it
+if [ ! -d /share/inadyn-cache ]; then
+    mkdir /share/inadyn-cache
+fi
+
 # Check it
 /usr/sbin/inadyn --check-config
 
-## If cache folder not exit then make
-#if [ ! -f /share/inadyn-cache ]; then
-#    mkdir /share/inadyn-cache
-#    chmod a+r /share/inadyn-cache
-#fi
-rm -rf /root/.cache/inadyn
-ln -s /share/inadyn-cache /root/.cache/inadyn
-#chmod a+rw /root/.cache/inadyn
+# If inadyn cache dir exists, check if it's not a symbolic link, if it's not a link then remove the folder and create a link
+# test -d is true if it's a dir or a link to a dir, so further needs to test if it's symbolic link
+if [ -d /root/.cache/inadyn ]; then
+    if [ ! -L /root/.cache/inadyn ]; then
+        rm -rf /root/.cache/inadyn
+        ln -s /share/inadyn-cache /root/.cache/inadyn
+    fi
+# If inadyn cache is false with test -d then it's false with test -L
+else
+    ln -s /share/inadyn-cache /root/.cache/inadyn
+fi
+
 
 ##############
 # Launch App #
